Support for Ajax calls added.
Note, that the current behaviour is to handle the ajax call and call the ajax callback handler as soon as a call is made. This is done by a small not-so-intuitive change in the XMLHttpRequest class in webkit. Furthermore, the necessary infrastructure has been put in place for detecting these calls and registering them in artemis, such that in the long run we can control when they are to be called from there (and again, code has been inserted into artemis which in principle does the same as done directly in the XMLHttpRequest class - handling the Ajax requests instantly). In the long run we should let Artemis have control over this.