Performance improvements for coverage
Usage of QUrl is replaced with QString, since QUrl has very poor performance. Furthermore, a QSource object is introduced, representing a unique source document. Over time, we should replace all usage of hashes and urls to identify source locations with a getID call on the QSource object (or a QCodeBlock object).