Partly fixed issues with trace visitors (see below).
However, this required me to replace the smart pointers with standard pointers in the visit() methods. When smart pointers are used, the node objects are destroyed once the smart pointer is no longer referenced (which is not what we want).

Is there a more robust way to pass a smart pointer to 'this' from the nodes' accept() methods?
Maybe this could be done with QSharedDataPointers, I have not worked out yet if that would be useful.

There may also be a similar problem with the smart pointers pointing to the visitor objects themselves, as these will be created and passed to accept() in the visit() methods once we implement some visitor which actually traverses the structure. If these pointers cause the visitors to be destroyed then we will be in trouble again.

Short of switching everything back to standard pointers, I'm not sure what exactly to do about this.