Debugging progress with TraceBuilder problem
I have tracked the segfault (at termination) down to the mTraceBuilder object in webkitexecutor. By commenting out its declaration and initialisation the program works, and leaving them in causes a segfault during destruction. I have stripped out all uses of this class and all methods and members from the class, so it is completely empty. I have no idea how this can be causing a problem.