Change path representation of path conditions
Now a path condition is represented using a list of tuples of of an symbolic expression and a boolan.

Each branch in a trace produces a new tuple. The symbolic expression represents the condition to take the "true" direction in the branch (jumps), and the boolean indicates if the trace followed this direction or not.