Fix issue with symbolic coverage highlighting in legend of coverage reports.
However, there aqre still issues with this highlighting. I think eventually (if we want to support incrementally showing/hiding different colours) we need to hve a system where lines have multiple annotations which are switched on and off by changing which class they belong to, not their individual style. This would allow the stylesheet to determine which should override each other and in what way automatically.

For example at the moment switching off a symCovered line reverts it to white, instead of to looking like it is simply covered. There are some other slightly strange interactions as well...