Add source indication to all symbolic values
A symbolic value now contains a "source" describing its origin. E.g. if the symbolic value flows from a input element, then this will be indicated including the name or id of the element.

Currently, only form inputs are supported and only id and name attributes are used to identify these. Thus, we assume that all input elements can be identified uniquely by one of these.

A warning will be printed if no name of id is given to a form element. However, if id or name is not unique the behaviour will be nondeterministic.