Fixed missing character issue with Kaluza's output.
The problem is the index used when refrmatting Kaluza's output into /tmp/kaluza-result, we accidentally remove a little more of each line than required.

It turns out the wawk manual includes " The first character of a string is character number one." in the substr() entry. However, every example I have ever seen (and done myself) uses 0 to refer to the beginning of the string, so I am amazed that this use could possibly have different behaviour on different systems. This fix works for me but I would be interested if it is also working for you Casper and Christian?