Work on the concolic mode.
This has added what turned out to be a very tricky part: matching up the FormFields on the page and the symbolic variables. Currently this is a huge hack, and we should discuss how this should really be dealt with. I suggest that there is some modification to SymbolicSource so that it can track the exact element which a given variable came from, or at least some easier way to match them up after-the-fact.

Also, it still does not run on the simple-validation-3 example, because the traces cannot be merged correctly. I will look into this now.