Add concolic test case for invisible fields.
Crerated this test case while investigating the issues with the alaska-air example. There we fail to inject into a field which is in a "folded-away" part of the form which is not visible to the user. I was suspicious that we may not handle the case of non-visible (but also non-hidden) form inputs correctly.