Add XPath entry-point selection to the demo mode.
For example, run
artemis --major-mode concolic http://www.rightmove.co.uk
and then click "set XPath for entry point" and try both "//input" and "id('buy')".

This is partly as a demo and partly in preparation/testing for adding the same feature to the concolic mode via the --concolic-button flag.