Initial work towards using an external entry-point finding tool.
In this version of the test suite, all the calls to the external tool are done at the start, and then the tests are run. I would prefer to avoid this and keep all processing of a single site together, but I also wanted to keep the individual test generation as simple as possible and not mix the calls to DIADEM and ArtForm too much.