Don't overwrite the onclick attribute on links if javascript: is used in the href
We overwrite the onclick attribute on links if they use the javascript: protocol, such that Artemis detect these as targets for executing event handlers.

However, some websites use both onclick and href="javascript:", the onclick event handler to validate input and the href to execute the action (if validation returns true). Thus, we should not overwrite the onclick if an onclick is already present.

Relates to #62