Correct typing of variables in the solver. [Issues reminaing]
However there are still issues. Because we don't really support understanding the variable types when reading back the solver's results we cannot interpret boolean values properly.

This is causing the boolean-coercion-of-string-to-false.html example to fail, because we inject "false" when in fact we want a string which evaluates to false (i.e. the empty string "").