Add new TraceConcreteSummarisation node type.
This node type si created instead of strings of TraceConcreteBranch and TraceFunctionCall nodes.

It does not yet support merging two traces where a concrete branch is explored in both directions. The trace merger is not yet updated to deal with the new node, so it will not detect any divergence within a concrete summary.