Fixed a bunch of bugs in the new search framework.
Now the random search procedure can solve all the concolic.py test cases. Some of the assertions do not pass because the conditions are explored in a different order so the final PC and final injection can be different. If the random procedure is modified to always return the first possibility (i.e. DFS) then all the tests pass.