Fix an issue with building the injection from the solver's solution.
The solution can contain values for variables which are not in the PC to begin with. This can happen (e.g.) when there is a radio button restriction involved. In that case, these variables MUST be incuded in the injection, otherwise we can inject something wihch does not satisfy the intended constraints.

Fixes a bug where we would inject all radio buttons with not-checked even though the solver knew to avoid this.