Concolic benchmarking: Replace hidden fields flag by a runtime flag for all concrete value properties.
That is inputs of type: button, hidden, submit, reset, image.

This is tricky because our feature of not making things symbolic before injection means that these fields cannot create a symbolic branch just by removing the check, as they are never added to the event list and so never marked as symbolic. To add this flag, we also have to inform Artemis that these inputs should be considered standard text boxes.