Add a new test suite for delegation tests.
This re-uses some functionality from the concolic.py tests, so that was refactored a bit.

So far the only tests reproduce those which already existed in tests.py under TargetSolverNoInjection which just check that a constraint could be read from the examples and solved successfully.

The next step is to add assertions to each test case to check they are handled correctly.