Update delegation test suite
The test suite is updated to include the assertions in each test page.

The status of the tests is as follows:

Test                             | Status
---------------------------------|--------------------
target-id-check                  | Working
target-multiple-access           | Artemis error due to bad constraint
dom-constraint-duplication       | Working
custom-similar-to-jquery-example | Working
prototype-on-tutorial            | Missed branch
jquery-1-11-context              | Artemis crash printing a symbolic condition
jquery-1-11-no-context           | Working (with CNS branches)
jquery-1-7                       | Working (with CNS branches)
jquery-on-demo                   | Missed branch

However there also seem to be some as-yet undiagnosed intermittent issues with any test, causing it to give an ArtemisCallException in the test suite.