Add a pixel-based user-visibility check for events.
Added a second implementation for checking if events are user-visible.
Instead of checking element bounding boxes (which has several corner cases which are not handled well) we can now inspect each pixel in the web view to see what element is under it. This is a very good model of which elements can be clicked. It has two downsides:
* It is EXTREMELY slow (performs layout checks over ~1M pixels).
* It can only check pixels visible in the web view. If the page extends outside the visible aea (currently 1200x800) then those parts will not be checked. We would need to implement some scrolling mechanism to overcome this.