Add support for symbolic null values being compared with strings in EQUAL and NOT_EQUAL conditions. Fixes #119.
It still remains to check if we need to add support for any other comparisons (e.g. STRICT_EQUAL), other types (e.g. Int == Null), or other bytecode operations or built-ins.

It may also be useful to add warnings for cases where we knowingly drop symbolic information in the symbolic interpreter, as was the case here.